media discoverer: compilation fix

+++ Sources/VLCMediaDiscoverer.m
@@ -29,6 +29,7 @@
 #import "VLCLibVLCBridging.h"
 #import "VLCEventManager.h"
 
+#include <vlc/vlc.h>
 #include <vlc/libvlc.h>
 #include <vlc/libvlc_media_discoverer.h>
 
@@ -57,7 +58,7 @@ NSString *const VLCMediaDiscovererCategory = @"VLCMediaDiscovererCategory";
 + (NSArray *)availableMediaDiscovererForCategoryType:(VLCMediaDiscovererCategoryType)categoryType
 {
     libvlc_media_discoverer_description_t **discoverers;
-    unsigned numberOfDiscoverers = libvlc_media_discoverer_list_get([VLCLibrary sharedInstance], (libvlc_media_discoverer_category_t)categoryType, &discoverers);
+    ssize_t numberOfDiscoverers = libvlc_media_discoverer_list_get([VLCLibrary sharedInstance], (libvlc_media_discoverer_category_t)categoryType, &discoverers);
 
     if (numberOfDiscoverers == 0) {
         libvlc_media_discoverer_list_release(discoverers, numberOfDiscoverers);
